% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecea.R
\name{doECEA}
\alias{doECEA}
\title{Performs Equivalent Change Enrichment Analysis}
\usage{
doECEA(
  gene_sets,
  eci,
  min_size = 10,
  max_size = 500,
  num_perm = 1000,
  fdr_cutoff = 0.2,
  seed = 1
)
}
\arguments{
\item{gene_sets}{list of gene sets, as in fgsea}

\item{eci}{Named vector of equivalent change indices from getECI()}

\item{min_size}{Minimum size of pathways}

\item{max_size}{Maximum size of pathways}

\item{num_perm}{Number of permutations for permutation test}

\item{fdr_cutoff}{FDR threshold for returned pathways}

\item{seed}{Random seed for permutation test}
}
\value{
data.table with all pathways and enrichment statistics
}
\description{
Performs Equivalent Change Enrichment Analysis
}
\examples{
ecea_res <- doECEA(reactome_sets, eci)

}
