% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecea.R
\name{getECI}
\alias{getECI}
\title{Generates equivalent change index, given two vectors of
effect sizes.}
\usage{
getECI(smd1, smd2, p1 = NULL, p2 = NULL)
}
\arguments{
\item{smd1}{Named vector of effect sizes for first treatment.}

\item{smd2}{Named vector of effect sizes for second treatment.}

\item{p1}{Named vector of unadjusted p-values for first treatment.}

\item{p2}{Named vector of unadjusted p-values for second treatment.}
}
\value{
named vector of ECI values
}
\description{
If you do not want to adjust the ECI by p-value, then
do not pass p-values.
}
\details{
NOTE: This function will subset the data to include only the
entries that are in common to both vectors.
}
\examples{
eci <- getECI(smd1, smd2, p1, p2)

}
