% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecea.R
\name{getReactome}
\alias{getReactome}
\title{Gets gene sets for use with ECEA.}
\usage{
getReactome(species = "human")
}
\arguments{
\item{species}{currently accepts 'human' or 'mouse'}
}
\description{
Gets gene sets for use with ECEA.
}
