% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecea.R
\name{getReactome}
\alias{getReactome}
\title{Gets gene sets for use with ECEA.}
\usage{
getReactome(species = "human", progress = TRUE)
}
\arguments{
\item{species}{character, currently accepts 'human' or 'mouse'}

\item{progress}{Boolean value indicating if progress bar should be shown.}
}
\description{
Gets gene sets for use with ECEA.
}
